Rename shapes/ to meshes/
