+++ Monkey/game.ts
@@ -1,7 +1,7 @@
 import {Material, Shape} from "../common/material.js";
 import {GL_CULL_FACE, GL_CW, GL_DEPTH_TEST} from "../common/webgl.js";
-import {mesh_monkey_flat} from "../shapes/monkey_flat.js";
-import {mesh_monkey_smooth} from "../shapes/monkey_smooth.js";
+import {mesh_monkey_flat} from "../meshes/monkey_flat.js";
+import {mesh_monkey_smooth} from "../meshes/monkey_smooth.js";
 import {Camera} from "./components/com_camera.js";
 import {Light} from "./components/com_light.js";
 import {start, stop} from "./core.js";
