+++ Shading/game.ts
@@ -1,8 +1,8 @@
 import {Material, Shape} from "../common/material.js";
 import {GL_CULL_FACE, GL_CW, GL_DEPTH_TEST} from "../common/webgl.js";
-import {mesh_cube} from "../shapes/cube.js";
-import {mesh_icosphere_flat} from "../shapes/icosphere_flat.js";
-import {mesh_icosphere_smooth} from "../shapes/icosphere_smooth.js";
+import {mesh_cube} from "../meshes/cube.js";
+import {mesh_icosphere_flat} from "../meshes/icosphere_flat.js";
+import {mesh_icosphere_smooth} from "../meshes/icosphere_smooth.js";
 import {Camera} from "./components/com_camera.js";
 import {Light} from "./components/com_light.js";
 import {start, stop} from "./core.js";
