+++ meshes/asset2shape.cjs
