+++ meshes/icosphere_flat.blend
