+++ meshes/icosphere_flat.ts
