+++ meshes/icosphere_smooth.ts
