+++ meshes/monkey_flat.blend
