+++ meshes/monkey_flat.ts
