+++ meshes/monkey_smooth.ts
